reorganize sale
