+++ schema/artist/index.js
@@ -6,10 +6,10 @@ import Image from '../image';
 import Article from '../article';
 import Artwork from '../artwork';
 import PartnerShow from '../partner_show';
-import Sale from '../sale';
+import Sale from '../sale/index';
 import ArtworkSorts from '../sorts/artwork_sorts';
 import PartnerShowSorts from '../sorts/partner_show_sorts';
-import SaleSorts from '../sorts/sale_sorts';
+import SaleSorts from '../sale/sorts';
 import ArtistCarousel from './carousel';
 import ArtistStatuses from './statuses';
 import gravity from '../../lib/loaders/gravity';
